% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{analyze}
\usage{
analyze(o)
}
\arguments{
\item{o}{An n-graph (n-by-n) matrix of 0s and 1s (aka adjacency matrix), as output by the ImitSim and BadApple
simulations in the object ObstyRecord}
}
\value{
{}{Returns an integer which is the connectivity of the n-graph.}
}
\description{
analyze
}
\examples{
analyze(obsty)
}
