% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effs_beta.R
\name{effs_beta}
\alias{effs_beta}
\title{effs_beta}
\usage{
effs_beta(betaobj, brange = c(0, 1), base, test)
}
\arguments{
\item{betaobj}{betamle object}

\item{brange}{vector}

\item{base}{vector}

\item{test}{vector}
}
\value{
vector
}
\description{
effs_beta
}
