% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_beta.R
\name{mean_beta}
\alias{mean_beta}
\title{mean_beta}
\usage{
mean_beta(betaobj, testvals, brange = c(0, 1))
}
\arguments{
\item{betaobj}{A simobj from the ImitSim() or BadApple() functions}
\item{testvals}{A vector length K (including constant) of values to compute a mean estimate.}
\item{brange}{A vector length 2 (min, max) specifying the range of the beta density}
}
\value{
\item{}{A K-length vector of the mean at the testvals}
}
\description{
mean_beta is used by effs_beta to compute first differences (changes in mean)
}
