% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{peek}
\usage{
peek(simobj, when)
}
\arguments{
\item{simobj}{A simulation object from BadApple() or ImitSim()}
\item{when}{An iteration}
}
\value{
}
\description{
peek() returns an ugly histogram of the performance of saboteurs and others. Requires ggplot be loaded. Really should
be renamed not to conflict with the base peek() function for handling queues.
}
\examples{
foo <- BadApple(10, NumSaboteurs=3)
peek(foo, 1)
}
