% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_ngraph.R
\name{plot_gg_ngraph}
\alias{plot_gg_ngraph}
\title{plot_gg_ngraph}
\usage{
plot_gg_ngraph(imit_obj, repl_num)
}
\arguments{
\item{imit_obj}{simulation objecct}

\item{repl_num}{integer}
}
\value{
ggplot object
}
\description{
plot_gg_ngraph
}
