% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearranger.R
\name{rearranger}
\alias{rearranger}
\title{Title}
\usage{
rearranger(simobj)
}
\arguments{
\item{simobj}{A simulation object from either ImitSim() or BadApple()}
}
\value{
\item{}{Returns a tibble with four columns: Replication, Iteration, Worker, Performance}
}
\description{
Used to return a tidyverse dataset from a simulation object documenting the performance of each worker by replication
and iteration. Dpes not require that tallperformance=TRUE be set.
}
\examples{
foo <- ImitSim(10)
boo <- rearranger(foo)
}
