% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitch_plots.R
\name{snitch_plots}
\alias{snitch_plots}
\title{snitch_plots}
\usage{
snitch_plots(simobj)
}
\arguments{
\item{simobj}{A simobject from BadApple}
}
\value{
\item{}{A list of ggplots}
}
\description{
snitch_plots produces a two-item list of ggplots (a histogram, a density plot).
}
\examples{
library(tidyverse)
foo <- BadApple(10, NumSaboteurs=3)
boo <- snitch_plots(foo)
# a histogram
boo[[1]]
# a density plot
}
