% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitch_tab.R
\name{snitch_tab}
\alias{snitch_tab}
\title{snitch_tab}
\usage{
snitch_tab(simobj)
}
\arguments{
\item{simobj}{A simulation object from BadApple(). Produces wrong values for ImitSim() because there are no designated saboteurs. Assumes NumSaboteurs=3.}
}
\value{
}
\description{
Produces a matrix with the differences between saboteurs and others by calling calc_saboteurs_vs_others(). Requires tidyverse.
}
