% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{analyze}
\usage{
analyze(obsty)
}
\arguments{
\item{obsty}{obsty is an n-by-n adjacency matrix, used in ImitSim and BadApple. Doesn't make much sense outside
of those functions. Calls igraph::vertex_connectivity.}
}
\value{

}
\description{
analyze
}
\examples{
}
