% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamle.r
\name{betamle}
\alias{betamle}
\title{Title}
\usage{
betamle(fmu, method = "BFGS", data, brange)
}
\arguments{
\item{fmu}{formula object}

\item{method}{character string}

\item{data}{dataset}

\item{brange}{vector}
}
\description{
Title
}
