% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamle.r
\name{betamle}
\alias{betamle}
\title{Title}
\usage{
betamle(formula, data, brange)
}
\arguments{
\item{formula}{Required, standard R formula.}
\item{brange}{Required, c(min,max) for range of the DV (inclusive)}
\item{data}{Required.}
}
\value{

}
\description{
Simple change to argument list (2:02pm, 14 july)
}
