% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamle.r
\name{betamle}
\alias{betamle}
\title{Maximum Likelihood Estimates of the Beta Density (2 parameter)}
\usage{
betamle(formula, data, brange)
}
\arguments{
\item{formula}{Required, standard R formula.}
\item{brange}{Required, c(min,max) for range of the DV (inclusive)}
\item{data}{Required.}
}
\value{
Returns nothing, just prints the output. I really do need to fix this.
}
\description{
Computes the MLE for the Beta density (2 parameter), defined over c(min,max). Computes coeficients for the $\nu, \omega$ with an exponential link, plus SEs, Z-scores, and p-values.

In a subsequent version, I would like the function to return as objects as well.
}
