% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_saboteurs_vs_others.R
\name{calc_saboteurs_vs_others}
\alias{calc_saboteurs_vs_others}
\title{calc_saboteurs_vs_others}
\usage{
calc_saboteurs_vs_others(simobj)
}
\arguments{
\item{simobj}{}
}
\value{

}
\description{
calc_saboteurs_vs_others
}
