% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effs_beta.R
\name{effs_beta}
\alias{effs_beta}
\title{effs_beta}
\usage{
effs_beta(betaobj, brange = c(0, 1), base, test)
}
\arguments{
\item{betaobj}{A simulation object from ImitSim() or BadApple()}
\item{brange}{A vector (min, max) specifying the range of the beta density}
\item{base}{A vector of length K (# vars + constant) for the baseline}
\item{test}{A vector of length K (# vars + constant) for the test values)}
}
\value{
\item{}{Returns a vector of changes in mean of the distribution given a change from base to test}
}
\description{
effs_beta calculates the mean shift of each variable as the variable moves from base to test conditions
}
}
