% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_beta.R
\name{mean_beta}
\alias{mean_beta}
\title{mean_beta}
\usage{
mean_beta(betaobj, testvals, brange = c(0, 1))
}
\arguments{
\item{betaobj}{betamle object}

\item{testvals}{vector}

\item{brange}{vector}
}
\value{
vector
}
\description{
mean_beta
}
