% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_by_iteration.R
\name{perf_by_iteration}
\alias{perf_by_iteration}
\title{perf_by_iteration()}
\usage{
perf_by_iteration(pobj, title = NA, iteration_list = NA, obj_type = "ImitSim")
}
\arguments{
\item{pobj}{simulation object}

\item{title}{string}

\item{iteration_list}{vector}

\item{obj_type}{character string}
}
\value{
ggplot object
}
\description{
perf_by_iteration()
}
