% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_by_iteration.R
\name{perf_by_iteration}
\alias{perf_by_iteration}
\title{perf_by_iteration()}
\usage{
perf_by_iteration(pobj, title = NA, iteration_list = NA, obj_type = "ImitSim")
}
\arguments{
\item{obj_type}{}
}
\value{

}
\description{
perf_by_iteration()
}
