% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bgraph.R
\name{plot_bgraph}
\alias{plot_bgraph}
\title{Tplot_bgraph}
\usage{
plot_bgraph(imit_obj, repl_num)
}
\arguments{
\item{imit_obj}{simulation object}

\item{repl_num}{integer}
}
\value{
plot object
}
\description{
Tplot_bgraph
}
