% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_ngraph.R
\name{plot_gg_ngraph}
\alias{plot_gg_ngraph}
\title{Plot an incomplete n-graph using ggraph}
\usage{
plot_ibgraph(imit_obj, repl_num)
}
\arguments{
\item{imit_obj}{An object from an imitation simmulation}
\item{repl_num}{A replication number}
}
\value{
\item{ggraph_obj}{Returns a \verb{ggraph} object}
}
\description{
Bloody hell. This should be in the damned file.

Takes as input the name of an output from a imitation simulation (\verb{imit_obj}) and a specific replication number, and then returns a \verb{ggraph} object. Can be combined with other calls to make a grid of n-graphs using \verb{gridExtra}.
}
