% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bigraph.R
\name{plot_gg_ngraph}
\alias{plot_gg_ngraph}
\title{(Incomplete) N-graph plotter using ggraph}
\usage{
plot_gg_ngraph(imit_obj, repl_num)
}
\arguments{
\item{imit_obj}{An object from the imitation models (so far, only ImitSim)}
\item{repl_num}{A replication number}
}
\value{
Returns a \verb{ggraph} object.
}
\description{
Produces a pretty, \verb{ggraph} of the incomplete n-graph of the network of bureaucrats found in the \verb{Obsty} matrix in the simulation object, \verb{imit_obj}. Uses \verb{ggraph} and \verb{igraph}, and probably best if these are loaded first. Can be arranged into multiple panes with the \verb{gridExtra::grid.arrange()} function.
}
