% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitch_plots.R
\name{plot_performance}
\alias{plot_performance}
\title{plot_performance}
\usage{
plot_performance(simobj)
}
\arguments{
\item{simobj}{Either an "ImitSim" or a "BadApple" simulation object}
\item{obj_type}{Either "ImitSim" (default) or "BadApple"}
\item{iteration_list}{An optional list of iterations; default is to print all}
\item{title}{A title for the plot}
}
\value{
}
\description{
plot_performance() plots a histogram of the mean performance at a given iteration.
}
\examples{
}
