% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitch_plots.R
\name{snitch_plots}
\alias{snitch_plots}
\title{snitch_plots}
\usage{
snitch_plots(simobj)
}
\arguments{
\item{simobj}{}
}
\value{
\item{plotlist}{A list with two ggplot() objects: a histogram of the saboteurs vs the others, and a kernel density plot of the saboteurs vs the others.}
}
\description{
snitch_plots
}
