% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitch_tab.R
\name{snitch_tab}
\alias{snitch_tab}
\title{snitch_tab}
\usage{
snitch_tab(simobj)
}
\arguments{
\item{simobj}{}
}
\value{
\item{rtable}{Returns the results of calc_saboteurs_vs_others in a knitr::kable()-ready tabular form.}
}
\description{
snitch_tab
}
